refactor(portal-framework-core, portal-plugin-dashboard): rename WidgetArea to GridWidgetArea and introduce FlexWidgetArea
- Renamed `WidgetArea` component to `GridWidgetArea` for clarity and specificity
- Introduced new `FlexWidgetArea` component for flexible layout needs
- Updated imports and usage across dependent files including storybook stories, tests, and route components
- Created `useWidgetArea` hook to encapsulate widget area logic
- Reorganized exports in index.ts to group related functionality
